servers/Retrieve/requirements.txt
asgi_correlation_id==4.3.4
fastapi==0.115.12
uvicorn==0.34.0
loguru==0.7.3
pyyaml==6.0.2
httpx==0.28.1
requests==2.32.3
biopython==1.85
openpyxl==3.1.5
openai==1.86.0
openai-agents==0.0.17
pandas==2.2.3
pymilvus==2.5.8
crawl4ai==0.7.0
aiohttp==3.11.18
beautifulsoup4==4.12.3
asyncio==3.4.3
tiktoken==0.9.0
fastapi-mcp==0.4.0
mcp>=1.13.0
python-dotenv
slowapi==0.1.9